backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pydantic==1.8.2
python-dotenv==0.19.0
python-multipart
telethon==1.28.5
cryptg==0.4.0
